// src/shared/types/generated/application/ItemScope.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ItemScope = { "kind": "all" } | { "kind": "inbox" } | { "kind": "trash" } | { "kind": "media_matches", item_id: number, } | { "kind": "recently_viewed" } | { "kind": "untagged" } | { "kind": "uncategorized" } | { "kind": "folder", folder_id: number, } | { "kind": "smart_folder", smart_folder_id: number, };